rpi/requirements.txt
Pygments==2.1.3
Twisted==16.3.0
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
ipython==5.0.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.5
ptyprocess==0.5.1
pygame==1.9.2b8
pyserial==3.1.1
pyzmq==15.3.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
txZMQ==0.7.4
wcwidth==0.1.7
wsgiref==0.1.2
zope.interface==4.2.0
